api/requirements.txt
dotenv==0.9.9
fastapi==0.135.2
langchain-community>=0.3.25
langchain-core>=1.2.8
langchain-groq==1.1.2
langchain-huggingface>=0.3.0
langchain-qdrant>=1.1.0
mysql-connector-python==9.1.0
pandas==2.3.0
python-dotenv==1.1.0
qdrant-client==1.17.1
sentence-transformers==4.1.0
uvicorn==0.42.0
